Re-export database schema
