+++ package.json
@@ -78,6 +78,7 @@
     "@inquirer/prompts": "7.0.1",
     "@paralleldrive/cuid2": "2.2.2",
     "@ronin/compiler": "0.12.3",
+    "@ronin/schema": "0.1.1",
     "chalk-template": "1.1.0",
     "get-port": "7.1.0",
     "ini": "5.0.0",
@@ -90,7 +91,6 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
-    "@ronin/schemas": "0.1.5",
     "@types/bun": "1.1.12",
     "@types/ini": "4.1.1",
     "@types/pluralize": "0.0.33",
